requirements.txt
langchain
langgraph
chromadb
sentence-transformers
google-generativeai
streamlit
fastapi
uvicorn
python-dotenv
google-genai